Add automatic testing for encoding
(And some formatting)

Testing for full line trailing whitespace is currently disabled, as the
working tree is full of it. This is easy enough enable again, if some
consensus can be reached on the matter.

My two cents is a preference to nix all kinds of trailing whitespace,
not just those that follow from content, but project consistency is
always most important.